+++ packages/app/public/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### Removed
+- Removed under construction messages.
+
 ## [3.20.0] - 2019-06-28
 ### Changed
 - Updated to Riot API version 9.13.1